vehicle/requirements.txt
numpy
scipy
# scikit-image
# tensorflow>=1.3.0
opencv-python
packaging
pillow
matplotlib
